+++ esp-rtos/src/scheduler.rs
@@ -363,7 +363,7 @@ impl SchedulerState {
         timer_queue.schedule_wakeup(task, at)
     }
 
-    #[esp_hal::ram]
+    #[ram]
     pub(crate) fn resume_task(&mut self, task: TaskPtr) {
         let timer_queue = unwrap!(self.time_driver.as_mut());
         timer_queue.timer_queue.remove(task);